refactor: move tabline code to statusline.c
Problem:	Tabline code is closely related to statusline, but still left over in drawscreen.c and screen.c.
Solution:	Move it to statusline.c(https://github.com/neovim/neovim/pull/20923#issuecomment-1305023156).
